refactor: 결정된 팀 컨벤션으로 수정
- 패키지 이름 변경하기(web, application, domain, infra)
- Fixture 위치 도메인 패키지로 옮기기
- 파라미터 final 붙이기
