+++ src/main/java/com/woowacourse/pelotonbackend/member/web/MemberController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.woowacourse.pelotonbackend.member.application.MemberService;
 import com.woowacourse.pelotonbackend.member.domain.Member;
-import com.woowacourse.pelotonbackend.member.service.MemberService;
 import com.woowacourse.pelotonbackend.member.web.dto.MemberRequest;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -21,7 +21,7 @@ public class MemberController {
     private final MemberService memberService;
 
     @PostMapping("/api/members")
-    public ResponseEntity<Void> createMember(@RequestBody @Valid MemberRequest memberRequest) {
+    public ResponseEntity<Void> createMember(@RequestBody @Valid final MemberRequest memberRequest) {
         final Member member = memberService.createMember(memberRequest.toMember());
         return ResponseEntity
             .created(URI.create("/api/members/" + member.getId()))
